chore(ci): move protobuf specs to repository top level (#629)
## Summary
Move the Astria protobuf specs into a top-level `proto/` folder and
perform Protobuf to Rust code generation in a self-standing compilation
tool.

## Background
The protobuf specifications are the single point of truth for Astria's
API. As such they should stand by themselves in the repository hierarchy
and not be a detail of the `astria-proto` Rust crate as a
language-specific implementation.

Furthermore, the current state using integration tests to compile
protobuf to rust code via integration tests made for bad developer
experience: a error compiler higher level abstractions on top of the
generated lower level interfaces lead to the test - and hence code
generation - never running. For example, this happened frequently due to
the programmer anticipating changes made to a protobuf message but
before re-generating code.

Protobuf-to-Rust compilation now happens in the
`astria-protobuf-compiler` binary crate, and the check for the Protobuf
spec and generate Rust code being in sync is moved to a github action.

## Changes
- Move `crates/astria-proto/proto` under the repo root to `proto/astria`
- Create an `astria-protobuf-compiler` tool that invokes `buf` and calls
`tonic-build`
- Run the protobuf compiler in CI and test if it leaves the git tree
unchanged, fail otherwise (as opposed to doing this check in Rust
before)

## Testing
All Rust code using the protobuf definitions still compiles. Deleting
the generated code prior to this change and then re-generating with the
new tool leaves the git tree unchanged. Deleting a single file makes the
github action fail (meaning the job checking for proto spec and Rust
being synced works).