fix(docker-build): amend Cargo.toml when building images (#672)
## Summary
Build images using a Cargo.toml without a default-members entry.

## Background
We use cargo-chef to prepare recipes for each individual binary so that
only its dependencies are built and cached. PR
https://github.com/astriaorg/astria/pull/664 introduced the
`default-members` entry so that cargo invocations by default don't
target the lints (as they require a nightly compiler). But this breaks
cargo-chef because it leaves default-members intact, leading to cargo
compilation issues.

## Changes
- Create a special `containerfiles/Cargo.toml` file that is derived from
the root `Cargo.toml` but with its `default-members` entry removed
- Update `containerfiles/Dockerfile` to run `cargo chef prepare` against
`containerfiles/Cargo.toml`
- Add a just entry to remove `workspace.default-members` from
`Cargo.toml`, format it with `taplo`, and write it to
`containerfiles/Cargo.toml`
- Add a github job that ensures that `containerfiles/Cargo.toml` is in
sync with `Cargo.toml`

## Testing
- Docker builds were broken and now work again (verified by building
locally and in CI)
- Commited a modified `containerfiles/Cargo.toml` and observed the
github action trigger

## Related Issues
PR causing the build to break:
https://github.com/astriaorg/astria/pull/664