refactor(conductor): use idiomatic types, streamline some executor code (#678)
## Summary
Defined idiomatic execution types in `astria-core`, refactored and
streamlined executor code.

## Background
The conversion method for commitment states would panic, it returns an
error now and the conversion happens in `astria-core`, similar to the
sequencer types.

There was also a lot of unused code in the execution service extension
trait, and unnecessary redirection when updating commitment states
(single-use functions that make reading the code opaque). All of that
has been removed.

## Changes
- Defined `Block` and `CommitmentState` in `astria-core` with conversion
methods from the raw generated prost types.
- Changed the exection service extension crate to a newtype wrapper on
`ExecutionServiceClient<Channel>` (it used to be that, but the newtype
wrapper fits better in this instance)
- Removed a few methods by inlining their logic at the call site.

## Testing
Executor tests still run, only refactoring.