feat(composer): bundle multiple rollup transactions into a single sequencer transaction (#651)
## Summary
This adds logic for bundling multiple `RollupTransaction`s into a single
sequencer `SignedTransaction`. It also introduces a `max_bundle_size`
cap to the amount of bytes that the `Executor` will submit.

The `Executor` now adds incoming `SequenceAction`s into the
`BundleFactory`, which manages a buffer of `SizedBundle`s waiting to be
submitted to the `Sequencer`. A `SizedBundle` is a `Vec<Action>` with a
limit on how many bytes can be packed into it.
The `Executor` submits a finished bundle as soon as an in-flight
submission returns a response.
If there are no finished bundles, the `Executor` will preempt the
current bundle being built within 1 sequencer block time of the last
submission response by forcing the `BundleFactory` to flush.

## Background
Product requirements of the `composer` service to be able to submit
bundles with more than a single `RollupTransaction`.

## Changes
- Add `SizedBundle`, `BundleFactory`, and `NextFinishedBundle` as well
as the logic driving it to the `Executor`
  - Added `max_bundle_sz` and `block_time` config vars
  - Added block timer logic to the `Executor`'s `select!` loop
- Removed `bundle_pending_tx` in the `Searcher` in favor of serializing
and forwarding a `SequenceAction` directly from the `Collector` to the
`Executor`
  - Removed `collector:RollupTransaction` since it is no longer needed
- Renamed `bundle_tx/rx` channel handles to `seq_actions_tx/rx` for
clarity
- `Searcher` will now only manage dead `Collector`s and `Executor`
states.
- Cleaned up some of the logging and spans in the `Executor`

## Testing
Added unit tests to the executor for the following:
- [x] Check that bundles are flushed when they go over capacity
- [x] Check that bundles are flushed when the block timer ticks
- [x] Check that two `SequenceAction`s are bundled into a single
`SignedTransaction` submission
Added unit tests to the bundling logic & associated structs:
- [x] `SizedBundle`
- [x] `BundleFactory` and `NextFinishedBundle`

## Breaking Changelist

## Related Issues
Link any issues that are related, prefer full github links.

closes #599
closes #598

---------

Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>