feat(ci): ensure all crate targets can be built outside the workspace (#919)
## Summary
Added the `--all-targets` flag to `cargo hack check` to ensure tests can
also be built outside the workspace.

## Background
While implementing https://github.com/astriaorg/astria/pull/889 it was
found that sequencer-relayer tests can only be run from the repository
root but not from the crate directory itself. The CI job that is
supposed to ensure this failed because it was not running with the right
flags.

## Changes
- Added `--all-targets` flag to `cargo hack check` in the rust test
github workflow.
- 
## Related Issues
Closes https://github.com/astriaorg/astria/issues/893