refactor(conductor)!: use HTTP to periodically poll Celestia heights  (#852)
## Summary
Replaces the websocket subscription to get the latest Celestia heights
by a HTTP fetch running periodically.

## Background
Websocket adds complexity for questionable gain. Tests in particular are
much more complex to set up, but also for normal operation disconnects
and resubscriptions need to be handled explicitly because websocket does
not have a reconnect mechanism.

This PR essentially comes full circle, removing every last trace of
websocket again.

## Changes
- Replace the websocket subscription by an interval stream + HTTP fetch
(not calling this GET because this is still JSONRPC dong a POST)
- Removes the `ASTRIA_CONDUCTOR_CELESTIA_NODE_WEBSOCKET_URL` config
- Adds the `ASTRIA_CONDUCTOR_CELESTIA_BLOCK_TIME_MS` config which takes
an integer representing the duration in milliseconds between two
successive Celestia block heights.
- Update charts to reflect the env var changes

## Testing
Will be tested using blackbox tests and smoke tests (out of scope of
this PR)

## Breaking Changelist
- Removed `ASTRIA_CONDUCTOR_CELESTIA_NODE_WEBSOCKET_URL`
- Added `ASTRIA_CONDUCTOR_CELESTIA_BLOCK_TIME_MS`