fix(sequecer): is_proposer check now considers proposer's address (#936)
## Summary

This PR makes the 'did I propose this' check in `process_proposal()`
more robust by caching the proposer's address in `prepare_proposal()`.
Changed the check from a bool to checking if the cached proposing
address matches the proposal's creator.

## Background
The sequencer's app logic shouldn't execute transactions in
`process_proposal` if it was the app that proposed the transactions in
`prepare_proposal` (as re-executing the transactions will fail due to
how we manage state). The previous check was only a `bool` that was set
in `process_proposal` and Zellic pointed out that this check could pass
incorrectly if the proposed proposal timed out and a new proposal was
proposed by someone else.

The proposing address is managed by cometbft and can
[change](https://github.com/cometbft/cometbft/blob/c3acea2436f4fc0d2cbbb8b27ada97ab3706892f/internal/consensus/state.go#L277).

This check will fail if there are multiple cometBFT validators running
with the same signing key.

I didn't check for proposal height because it seemed redundant; assuming
that we're not running multiple validators with the same key, if
`process_proposal` is called and we proposed it we must've generated it
in `prepare_proposal`.

## Testing
Ran locally and transaction submitting still works.

## Related Issues
closes #920