feat(composer): add graceful shutdown (#854)
## Summary
Implements graceful shutdown for Composer components. 

## Background
Composer has 3 main components:
1. Geth Collectors
2. Grpc Server/Collector
3. Executor

Currently, we do not have any sort of graceful shutdown for these
Components. This PR implements it.

When a Geth Collector shuts down, we do the following:
1. We need to unsubscribe to the websocket stream
2. We need to set the status of the geth collector to false(not
connected)

For the Grpc Server, we just pass in a future which resolves when a
shutdown signal is passed to the server by running it as
`serve_with_incoming_shutdown`.

When an Executor shuts down, we do the following:
1. Close the `serialized_rollup_transactions` channel to avoid receiving
transactions.
2. Set Executor status to false(not connected).
3. Drain any remaining bundles

We ensure that the API server shuts down at the end by giving it its own
CancellationToken. This is because we need the API server to be up in
order for k8s to report liveness.

## Changes
- Add a shutdown token which is a tokio CancellationToken
- Pass it to the GethCollectors, GrpcServer and Executor
- Convert the while loop which listens to txs on the Geth Wss stream to
a select stmt in a loop to listen for the shutdown signal from the
shutdown CancellationToken.
- Add shutdown logic to Executor
- Pass in the shutdown CancellationToken to the gRPC server with the
`serve_with_incoming_shutdown` method.
- Pass a new CancellationToken to the API server and ensure it shuts
down after every other task is shutdown.

**Note:** 
I had to refactor the run_until_stopped method in Composer as Clippy was
complaining a lot about the below:

https://rust-lang.github.io/rust-clippy/master/index.html#/too_many_lines

## Testing
Tested by manually deploying a local shared sequencer cluster and
running a Composer instance against it. Sent a SIGTERM signal to
Composer and tested the shutdown.