chore: update codeowners w/ teams (#934)
## Summary
Updating the split up into more groups. I have updated the groups within
github to autoassign a member from the group when added in a round robin
fashion. So this will auto assign, and if one person is in multiple
assigned groups selected that will count for all. I've also removed
myself from the assignment pool queue for rust reviewers.

## Background
We want to declutter inbox and more more concrete assignment. 

## Changes
- More granular code ownership settings, pushing most to groups.