refactor(composer): Use config-like Builders instead of `new` Constructors (#953)
## Summary
Add config-like Builders for Executor, GethCollector and GrpcServer.

## Background
Currently, Executor, GethCollector and GrpcServer components are
initialized by a `new()` constructor. As we keep adding more features,
the variables passed to the constructor can increase which can harm
readability.

This PR aims to increase readability by maintaining seperate Builders
for each component.

## Changes
- Create a `Builder` for `Executor` crate
- Create a `GethCollectorBuilder` for `GethCollector` struct
- Create a `GrpcServerBuilder` for `GrpcServer`.
- Remove `new()` method from Executor, GethCollector and GrpcServer

## Testing
Since this is a refactor and not a functional change, I ran the unit
tests and blackbox successfully which should be fine.